migrate functionality to utilities, add class mapping utilities
