+++ net/utilities/utils.py
@@ -1,48 +1,58 @@
 import torch
 from torch import nn
 import torchvision.models as models
+import matplotlib.pyplot as plt
+import numpy as np
 
-'''utils.py: Custom network object utilities '''
+'''utils.py: General Utilities '''
 __author__ = "Luis Quinones"
 __email__ = "luis@complicitmatter.com"
 __status__ = "Prototype"
 
-class net_utilities(object):
+class Utilities(object):
     '''
-    Basic static utilities class for net operations. 
+    Basic static utilities class. 
 
     '''    
 
-    @staticmethod    
-    def get_pretrained_model(name = 'vgg16', trained = True):
-        '''Generates the nn.module container Sequential classfier as the default for this class.
+    @staticmethod
+    def map_idx_to_classes(class_to_idx, classes, predicted_idx, predicted_prob):
+        '''
+        Maps the predicted indexes to the keys which we need to retrieve the class names. Since our model gives us the 'value',
+        we need to find the key for our class_to_idx dict, once we have the key we can use it to find the class mapping 
+        (flower name in this case).
 
         Args:
-            name (str): The pretrained model name ('vgg16', 'resnet50', 'densenet121').
-            trained (bool): If the model has been trained.
+            class_to_idx (dic of ints): This is where we store the dictionary mapping the name of the class to the index (label).
+            classes (dict of strs): Dict containing the mapping of the class idx to the name.
+            predicted_idx (list of ints): The topk list of predicted indexes.
+            predicted_prob (list of floats): The probability list from topk.
 
         Raises:
             TODO: Update exceptions with error_handling class.
 
         Returns:
-            model (torchvision.models.vgg.VGG): The torch vision model specified
-        '''        
+            idx_classes_dict (dict): Dictionary containing 'predicted_indexes': indexes predicted by network, 
+                                                            'idx_to_class': mapped idx_to_class, 
+                                                            'classes': class names,
+                                                            'probabilites': the probabilities for classes.
 
-        # get model from torchvision
-        if name == 'vgg16':
-            model = models.vgg16(pretrained = trained)
-        elif name == 'resnet50':
-            model = models.resnet50(pretrained = trained)
-        elif name == 'densenet121':
-            model = models.densenet121(pretrained = trained)
-        else:
-            raise ValueError('Please select from either vgg16, resnet50 or \
-                            densenet121 pre-trained models')
-            
-        # freeze parameters
-        for parameter in model.parameters():
-            parameter.requires_grad = False
+        '''
 
-        return model  
+        idx_classes_dict = {}
+        predicted_class_names = []
+        predicted_idx_to_class = []
 
+        for x in predicted_idx:
+            for k,v in class_to_idx.items():
+                if x == v:
+                    predicted_class_names.append(classes[k])
+                    predicted_idx_to_class.append(k)
+
+        idx_classes_dict['predicted_idx'] = predicted_idx
+        idx_classes_dict['classes'] = predicted_class_names
+        idx_classes_dict['idx_to_class'] = predicted_idx_to_class
+        idx_classes_dict['probabilities'] = predicted_prob
+
+        return idx_classes_dict
